Subject: Story 1645 - cfgd locate saved config on boot
Changed tab to spaces.

Added support to cfgd to read the configdb and find type=startup row. Cfgd
does NOT yet push the configuration to OVSDB since it requires a python
module that has not yet been created.

Also changed cfgd to NOT try to locate and push a configuration if the value of
open_vswitch:cur_cfg > 0.  This is to prevent the running configuration from
being overwritten by a saved configuration by someone starting up cfgd.

Change-Id: I5b105addde342b2b4e4aa431f196339944844f60
Signed-off-by: Gary Evans <gary.evans@hp.com>